requirements.txt
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
appnope==0.1.2
argon2-cffi==20.1.0
arrow==1.0.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
binaryornot==0.4.4
bleach==3.3.0
blessings==1.7
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
cookiecutter==1.7.2
dask==2021.4.0
decorator==5.0.5
defusedxml==0.7.1
entrypoints==0.3
filelock==3.0.12
fsspec==0.9.0
google-api-core==1.26.3
google-auth==1.28.1
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio==1.37.0
hiredis==2.0.0
idna==2.10
ipykernel==5.5.3
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
locket==0.2.1
MarkupSafe==1.1.1
mistune==0.8.4
modin==0.9.1
msgpack==1.0.2
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy==1.19.5
nvidia-ml-py3==7.352.0
opencensus==0.7.12
opencensus-context==0.1.2
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.8.2
partd==1.2.0
pexpect==4.8.0
pickleshare==0.7.5
poyo==0.5.0
prometheus-client==0.10.0
prompt-toolkit==3.0.18
protobuf==3.15.8
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.5
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.0
python-slugify==4.0.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
ray==1.1.0
redis==3.5.3
requests==2.25.1
rsa==4.7.2
Send2Trash==1.5.0
six==1.15.0
-e git+https://github.com/achung3071/Spring21-Capstone.git@a272d3185c4da9a0f41bc2bd506357f6d10ad455#egg=src
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.6.3
